MDEV-13073 Part 2 of merge of Ali semisync performance patch
Integration is completed to a level of no crashes but with few tests
reporting 'result mismatch' or 'warnings in the error log'.
Also PSI may not be fully merged and neither its test were run.

This patch refines control over the plugin hooks. When semisync is enabled
the "dynamic" hooks of plugins won't be run.

The total patch should be fine to try benchmarking while in parallel
its limitation will be lifted and possible merge glitches removed.

Todo: investigate properly

  There were few bugfixes found that are present in mysql and its unclear
  whether/how they are covered. Those include:

  Bug#15985893: GTID SKIPPED EVENTS ON MASTER CAUSE SEMI SYNC TIME-OUTS
  Bug#17932935 CALLING IS_SEMI_SYNC_SLAVE() IN EACH FUNCTION CALL
                 HAS BAD PERFORMANCE
  Bug#20574628: SEMI-SYNC REPLICATION PERFORMANCE DEGRADES WITH A HIGH NUMBER OF THREADS